+++ CHANGELOG.md
@@ -3,27 +3,66 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.1](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0...v1.0.1) (2019-10-29)
+# [1.0.0-beta.7](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-beta.4@beta...v1.0.0-beta.7@beta) (2019-11-06)
+
+
+### Changes
+
+* refactor Key Vault references ([#88](https://github.com/jazida-opensource/dotenv-azure/pull/88)) ([f6d1a65](https://github.com/jazida-opensource/dotenv-azure/commit/f6d1a65))
+
+# [1.0.0-beta.4](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-beta.3@beta...v1.0.0-beta.4@beta) (2019-10-14)
 
 
 ### Bug Fixes
 
-* **readme.md:** fix code examples ([#67](https://github.com/jazida-opensource/dotenv-azure/issues/67)) ([119fd23](https://github.com/jazida-opensource/dotenv-azure/commit/119fd23))
+* rate limit requsts by second ([#45](https://github.com/jazida-opensource/dotenv-azure/issues/45)) ([5cb9de3](https://github.com/jazida-opensource/dotenv-azure/commit/5cb9de3))
 
-# 1.0.0 (2019-09-19)
+# [1.0.0-beta.3](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-beta.2@beta...v1.0.0-beta.3@beta) (2019-10-04)
+
+
+### Bug Fixes
+
+* bug fixes ([31a2669](https://github.com/jazida-opensource/dotenv-azure/commit/31a2669))
+* limit concurrency to avoid Azure AD rate limiting ([#29](https://github.com/jazida-opensource/dotenv-azure/issues/29)) ([c5b2aeb](https://github.com/jazida-opensource/dotenv-azure/commit/c5b2aeb))
+* remove console.log ([5efa352](https://github.com/jazida-opensource/dotenv-azure/commit/5efa352))
+* semantic-release version ([5d355b2](https://github.com/jazida-opensource/dotenv-azure/commit/5d355b2))
+* tests, coverage and changelog ([74c0679](https://github.com/jazida-opensource/dotenv-azure/commit/74c0679))
+* update semantic-release dependencies ([7e1c970](https://github.com/jazida-opensource/dotenv-azure/commit/7e1c970))
+* version ([9e54cfd](https://github.com/jazida-opensource/dotenv-azure/commit/9e54cfd))
+
+# [1.0.0-beta.2](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-beta.1@beta...v1.0.0-beta.2@beta) (2019-10-04)
+
+
+### Bug Fixes
+
+* limit concurrency to avoid Azure AD rate limiting ([#29](https://github.com/jazida-opensource/dotenv-azure/issues/29)) ([c5b2aeb](https://github.com/jazida-opensource/dotenv-azure/commit/c5b2aeb))
+
+# 1.0.0-beta.1 (2019-10-02)
 
 
 ### Bug Fixes
 
 * don't throw on preload ([2abb93b](https://github.com/jazida-opensource/dotenv-azure/commit/2abb93b))
 * include config files to support preload ([5785633](https://github.com/jazida-opensource/dotenv-azure/commit/5785633))
-* include config files to support preload ([f1d92aa](https://github.com/jazida-opensource/dotenv-azure/commit/f1d92aa))
 
 
 ### Features
 
 * add preload ([223c210](https://github.com/jazida-opensource/dotenv-azure/commit/223c210))
-* add preload ([5ba742e](https://github.com/jazida-opensource/dotenv-azure/commit/5ba742e))
+
+# [1.0.0-alpha.5](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-alpha.4@alpha...v1.0.0-alpha.5@alpha) (2019-09-19)
+
+
+### Bug Fixes
+
+* tests, coverage and changelog ([74c0679](https://github.com/jazida-opensource/dotenv-azure/commit/74c0679))
+
+# [1.0.0-alpha.4](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-alpha.3@alpha...v1.0.0-alpha.4@alpha) (2019-09-19)
+
+
+### Bug Fixes
+
+* bug fixes ([31a2669](https://github.com/jazida-opensource/dotenv-azure/commit/31a2669))
 
 # [1.0.0-alpha.3](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-alpha.2@alpha...v1.0.0-alpha.3@alpha) (2019-09-16)
 
@@ -31,12 +70,6 @@ All notable changes to this project will be documented in this file. See
 ### Bug Fixes
 
 * don't throw on preload ([2abb93b](https://github.com/jazida-opensource/dotenv-azure/commit/2abb93b))
-* include config files to support preload ([5785633](https://github.com/jazida-opensource/dotenv-azure/commit/5785633))
-
-
-### Features
-
-* add preload ([223c210](https://github.com/jazida-opensource/dotenv-azure/commit/223c210))
 
 # [1.0.0-alpha.2](https://github.com/jazida-opensource/dotenv-azure/compare/v1.0.0-alpha.1@alpha...v1.0.0-alpha.2@alpha) (2019-09-16)